Remove FEED_URL Input
FEED_URL is an input but that variable isn’t referenced later. 
Also, the Input makes it appear in the override so if the parent recipe ever changes the yaml path, the override won’t get the change as the original value is still in the override recipe.  I suggest taking the FEED_URL out all together and just let the download recipe have that value hard coded like it is already.