Fixing Uninstall Script
Previously we only had one Woodwing product installed, which caused the uninstallation to delete the "Woodwing" plugin folder.  This masked an error when revising the "find" command which caused "find" to fail completely.  Now that we have Elvis/Assets in play, edit this correctly to completely remove the offending line rather than comment it out.