Deprecating Wipster Review Panel Recipes
Since it appears the direct download for Wipster Review Panel here is now gone in a week-old edit of the support article:
https://intercom.help/wipster-support/en/articles/3350755-where-do-i-get-the-wipster-review-panel

And two requests to get it reinstated have been ignored:
https://twitter.com/foigus/status/1461667653471375362
https://twitter.com/foigus/status/1466768852747005955

Deprecating these recipes.  My GenericZXPRepackage series of recipes might work if one could obtain an unencrypted ZXP format version of the plugin, but since I no longer need this recipe I'm not going to attempt to determine if this is the case.